+++ apps/epistola/src/main/kotlin/app/epistola/suite/templates/DocumentTemplateRoutes.kt
@@ -14,6 +14,7 @@ class DocumentTemplateRoutes(private val handler: DocumentTemplateHandler) {
             GET("", handler::list)
             GET("/search", handler::search)
             POST("", handler::create)
+            GET("/{id}/edit", handler::edit)
         }
     }
 }